refactor(events): move the SQLSpec sink into the SQLSpec backend and name modules for their contents
Two event modules were named after the libraries they integrate with, inside a
package where those names are also top-level imports. The SQLSpec sink is
backend-specific and now sits beside the SQLSpec event log; the Channels sink
keeps its home in events and says what it holds.

The WebSocket and SSE stream internals moved to the streaming module that was
already their only caller, so the Channels module holds only its sink.
